fixed startup times
